requirements.txt
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.19
boto==2.36.0
celery==3.1.17
croniter==0.3.5
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.4
django-admin-shortcuts==1.2.5
django-adv-imagefield==0.0.4
django-allauth==0.19.1
django-appconf==0.6
django-bootstrap-form==3.1
django-cacheops==2.3
django-debug-toolbar==1.2
django-endless-pagination==2.0
django-extensions==1.5.0
django-filebrowser-no-grappelli-nph==3.4.4
django-haystack==2.3.1
django-haystack-panel==0.2.1
django-hunger==2.1.2
django-jsonfield==0.9.13
django-location-field==1.5.1
django-markdown-deux==1.0.5
django-model-utils==2.0.3
django-pagedown==0.1.0
django-phonenumber-field==0.7.1
django-pipeline==1.4.6
django-redis==3.8.2
django-split-json-widget==0.2
django-storages-redux==1.2.2
django-tastypie==0.12.1
django-toolbelt==0.0.1
djangocms-admin-style==0.2.5
djrill==1.3.0
ecdsa==0.11
elasticsearch==1.3.0
Fabric==1.10.0
funcy==1.4
futures==2.2.0
gnureadline==6.3.3
gunicorn==19.1.1
ipython==2.2.0
jsonfield==1.0.0
kombu==3.0.24
markdown2==2.3.0
newrelic==2.40.0.34
oauthlib==0.7.2
paramiko==1.15.1
phonenumbers==7.0.2
Pillow==2.7.0
PredictionIO==0.8.3
psycopg2==2.5.2
pycrypto==2.6.1
python-dateutil==2.2
python-memcached==1.53
python-mimeparse==0.1.4
python-openid==2.2.5
pytz==2014.10
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
simplejson==3.4.1
six==1.9.0
sorl-thumbnail==12.2
sqlparse==0.1.12
static3==0.5.1
urllib3==1.10
